Cache fixes

+++ src/IBWT.Framework/State/Providers/IStateCacheProvider.cs
@@ -2,8 +2,8 @@ namespace IBWT.Framework.State.Providers
 {
     public interface IStateProvider
     {
-        StateContext GetState(int id);
-        bool Clear(int id);
+        StateContext GetState(long id);
+        bool Clear(long id);
         void ClearAll();
     }
 }
